+++ pyproject.toml
@@ -32,7 +32,6 @@ python = "^3.10,<3.13"
 sdsstools = "^1.2.1"
 sdss-clu = "^2.1.2"
 click-default-group = "^1.2.2"
-setuptools = "^67.6.0"
 typing-extensions = "^4.5.0"
 httpx = "^0.24.1"
 kubernetes = "^26.1.0"
