+++ src/gort/tools.py
@@ -10,6 +10,7 @@ from __future__ import annotations
 
 import asyncio
 import concurrent.futures
+import datetime
 import functools
 import hashlib
 import os
@@ -23,7 +24,10 @@ from functools import partial
 from typing import TYPE_CHECKING, Any, Callable, Coroutine, Sequence
 
 import httpx
+import numpy
+import pandas
 import peewee
+import pyarrow
 from astropy import units as uu
 from astropy.coordinates import angular_separation as astropy_angular_separation
 
@@ -64,6 +68,7 @@ __all__ = [
     "get_md5sum",
     "mark_exposure_bad",
     "handle_signals",
+    "GuiderMonitor",
 ]
 
 AnyPath = str | os.PathLike
@@ -734,3 +739,187 @@ def get_by_source_id(source_id: int) -> dict | None:
         return None
 
     return data[0]
+
+
+class GuiderMonitor:
+    """A tool to monitor guider outputs and store them in a dataframe."""
+
+    def __init__(self, gort: GortClient, actor: str | None = None):
+
+        self.gort = gort
+        self.actor = actor
+
+        self.data: pandas.DataFrame | None = None
+
+        self.__dtypes = {
+            "frameno": "int32[pyarrow]",
+            "telescope": "string[pyarrow]",
+            "time": pandas.ArrowDtype(pyarrow.timestamp("ms", "UTC")),
+            "n_sources": "int32[pyarrow]",
+            "focus_position": "float32[pyarrow]",
+            "fwhm": "float32[pyarrow]",
+            "ra": "float64[pyarrow]",
+            "dec": "float64[pyarrow]",
+            "ra_offset": "float32[pyarrow]",
+            "dec_offset": "float32[pyarrow]",
+            "separation": "float32[pyarrow]",
+            "pa": "float32[pyarrow]",
+            "pa_offset": "float32[pyarrow]",
+            "zero_point": "float32[pyarrow]",
+            "mode": "string[pyarrow]",
+            "ax0_applied": "float32[pyarrow]",
+            "ax1_applied": "float32[pyarrow]",
+            "rot_applied": "float32[pyarrow]",
+        }
+        self._data: dict[str, dict[tuple[int, str], Any]] = {
+            column: {}
+            for column in self.__dtypes
+            if column not in ["frameno", "telescope"]
+        }
+
+    def reset(self):
+        """Resets the internal state."""
+
+        self.__init__(self.gort, self.actor)
+
+    def start_monitoring(self):
+        """Starts monitoring the guider outputs."""
+
+        if self._handle_guider_reply not in self.gort._callbacks:
+            self.gort.add_reply_callback(self._handle_guider_reply)
+
+        self.reset()
+
+    def stop_monitoring(self):
+        """Stops monitoring the guider outputs."""
+
+        if self._handle_guider_reply in self.gort._callbacks:
+            self.gort.remove_reply_callback(self._handle_guider_reply)
+
+    def __del__(self):
+        self.stop_monitoring()
+
+    def update(self):
+        """Updates the internal dataframe."""
+
+        series: dict[str, pandas.Series] = {}
+        for column in self._data:
+            series[column] = pandas.Series(
+                list(self._data[column].values()),
+                index=self._data[column],
+                dtype=self.__dtypes[column],
+            )
+
+        df = pandas.DataFrame(series)
+        df.index.names = ["frameno", "telescope"]
+        df.reset_index(inplace=True)
+
+        df.frameno = df.frameno.astype("int32[pyarrow]")
+        df.telescope = df.telescope.astype("string[pyarrow]")
+        df.set_index(["telescope", "frameno"], inplace=True)
+        df.sort_index(inplace=True)
+
+        self.data = df
+
+        return df
+
+    async def _handle_guider_reply(self, reply: AMQPReply):
+        """Processes an actor reply and stores the collected data."""
+
+        if self.actor is not None:
+            if self.actor not in str(reply.sender):
+                return
+        else:
+            if ".guider" not in str(reply.sender):
+                return
+
+        body = reply.body
+
+        telescope = str(reply.sender).split(".")[1]
+        frameno: int | None = None
+        new_data: dict[str, Any] = {}
+
+        try:
+            if "frame" in body:
+                frame = body["frame"]
+                frameno = frame["seqno"]
+                new_data = {
+                    "time": pandas.to_datetime(datetime.datetime.now()),
+                    "n_sources": frame["n_sources"],
+                    "focus_position": frame["focus_position"],
+                    "fwhm": frame["fwhm"],
+                }
+
+            elif "measured_pointing" in body:
+                measured_pointing = body["measured_pointing"]
+                frameno = measured_pointing["frameno"]
+                new_data = {
+                    "ra": measured_pointing["ra"],
+                    "dec": measured_pointing["dec"],
+                    "ra_offset": measured_pointing["radec_offset"][0],
+                    "dec_offset": measured_pointing["radec_offset"][1],
+                    "separation": measured_pointing["separation"],
+                    "pa": measured_pointing.get("pa", numpy.nan),
+                    "pa_offset": measured_pointing.get("pa_offset", numpy.nan),
+                    "zero_point": measured_pointing.get("zero_point", numpy.nan),
+                    "mode": measured_pointing["mode"],
+                }
+
+            elif "correction_applied" in body:
+                correction_applied = body["correction_applied"]
+                frameno = correction_applied["frameno"]
+                new_data = {
+                    "ax0_applied": correction_applied["motax_applied"][0],
+                    "ax1_applied": correction_applied["motax_applied"][1],
+                    "rot_applied": correction_applied.get("rot_applied", 0.0),
+                }
+            else:
+                return
+
+            if not isinstance(frameno, int):
+                return
+
+            index = (frameno, telescope)
+            for column, value in new_data.items():
+                self._data[column][index] = value
+
+            self.update()
+
+        except Exception as err:
+            self.gort.log.warning(f"Error processing guider reply: {err}")
+
+    def to_header(self):
+        """Returns a header with pointing and guiding information."""
+
+        header: dict[str, Any] = {}
+
+        telescopes = ["sci", "spec", "skye", "skyw"]
+        if self.actor is not None:
+            telescopes = [self.actor.split(".")[1]]
+
+        if self.data is not None:
+            for tel in telescopes:
+                try:
+                    tel_data = self.data.loc[tel].copy().reset_index()
+                    if len(tel_data) < 2:
+                        frame0 = None
+                        framen = None
+                    else:
+                        frame0 = int(tel_data.frameno.min())
+                        framen = int(tel_data.frameno.max())
+
+                    header.update(
+                        {
+                            f"G{tel.upper()}FR0": (frame0, f"{tel} first guider frame"),
+                            f"G{tel.upper()}FRN": (framen, f"{tel} last guider frame"),
+                        }
+                    )
+
+                except Exception as err:
+                    self.gort.specs.write_to_log(
+                        f"Failed updating guider header information for {tel}: {err}",
+                        "warning",
+                    )
+                    continue
+
+        return header
